deployment/requirements.txt
streamlit
numpy
pandas
keras
scikit-learn
pillow
tensorflow-cpu